refactor(theme): remove redundant text styling now handled by theme
